requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
feedparser>=6.0.11
playwright>=1.44.0
anthropic>=0.28.0
supabase>=2.5.0
apscheduler>=3.10.4
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
pytz>=2024.1
certifi>=2024.0.0
